+++ noodles-tabix/src/async.rs
@@ -1,13 +1,19 @@
-//! Async tabix index and fields.
+//! Async tabix.
 
-mod reader;
+pub mod io;
 mod writer;
 
-pub use self::{reader::Reader, writer::Writer};
+#[deprecated(
+    since = "0.45.0",
+    note = "Use `noodles_tabix::r#async::io::Reader` instead."
+)]
+pub use self::io::Reader;
+
+pub use self::writer::Writer;
 
 use std::path::Path;
 
-use tokio::{fs::File, io};
+use tokio::fs::File;
 
 use super::Index;
 
@@ -26,7 +32,7 @@ use super::Index;
 /// # Ok(())
 /// # }
 /// ```
-pub async fn read<P>(src: P) -> io::Result<Index>
+pub async fn read<P>(src: P) -> tokio::io::Result<Index>
 where
     P: AsRef<Path>,
 {
@@ -52,7 +58,7 @@ where
 /// # Ok(())
 /// # }
 /// ```
-pub async fn write<P>(dst: P, index: &Index) -> io::Result<()>
+pub async fn write<P>(dst: P, index: &Index) -> tokio::io::Result<()>
 where
     P: AsRef<Path>,
 {