+++ noodles-tabix/src/async/io.rs
@@ -1,0 +1,5 @@
+//! Async tabix I/O.
+
+mod reader;
+
+pub use self::reader::Reader;