+++ noodles-tabix/src/lib.rs
@@ -38,7 +38,14 @@ pub use self::io::Reader;
 pub use self::io::Writer;
 
 #[cfg(feature = "async")]
-pub use self::r#async::{Reader as AsyncReader, Writer as AsyncWriter};
+#[deprecated(
+    since = "0.45.0",
+    note = "Use `noodles_tabix::r#async::io::Reader` instead."
+)]
+pub use self::r#async::Reader as AsyncReader;
+
+#[cfg(feature = "async")]
+pub use self::r#async::Writer as AsyncWriter;
 
 use std::{fs::File, path::Path};