+++ noodles-tabix/CHANGELOG.md
@@ -6,11 +6,17 @@
 
   * tabix: Move reader (`Reader`) and writer (`Writer`) to `io` module.
 
+  * tabix/async: Move reader (`Reader`) to `io` module.
+
 ### Deprecated
 
   * tabix: Deprecate `Reader` and `Writer`.
 
     Use `tabix::io::Reader` and `tabix::io::Writer`, respectively, instead.
+
+  * tabix: Deprecate `AsyncReader`.
+
+    Use `tabix::r#async::io::Reader` instead.
 
 ## 0.44.0 - 2024-09-04